pyproject.toml
[project]
name = "tpa"
version = "0.1.0"
description = "Terminal Pixel Art - An Aseprite-inspired pixel art editor for the terminal"
requires-python = ">=3.9"
dependencies = [
    "pillow>=11.3.0",
]

[project.optional-dependencies]
png = ["Pillow>=9.0"]

[project.scripts]
tpa = "tpa.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
